requirements.txt
audioop-lts
pillow
python-dotenv
py-cord
websocket-client